+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 # Change Log
 
-## v0.0.x Unreleased
+## v0.1.1
+
+### Maintenance and fixes
+
+* Fix group-specific effects matrix shape (#35)
+
+### Documentation
+
+## v0.1.0
 
 ### New features
+
+* Factor of group specific effect can be an interaction now (#31)
 
 ### Maintenance and fixes
 
 * Full interaction does not result in shared components anymore (#30)
+* Add information about levels to interaction terms (#33)
 
 ### Deprecation